requirements.txt
gunicorn==21.2.0
Flask==2.0.1
Pillow==8.4.0
numpy==1.21.5
